move token fabric method to Token class
