+++ api/src/main/java/org/accula/api/auth/jwt/refresh/RefreshTokenException.java
@@ -2,6 +2,9 @@ package org.accula.api.auth.jwt.refresh;
 
 import lombok.Getter;
 
+/**
+ * @author Anton Lamtev
+ */
 @Getter
 final class RefreshTokenException extends RuntimeException {
     private static final long serialVersionUID = -1081904770207522419L;
