+++ api/src/main/java/org/accula/api/detector/SuffixTreeUtils.java
@@ -3,7 +3,6 @@ package org.accula.api.detector;
 import com.google.common.collect.Streams;
 import com.suhininalex.suffixtree.Edge;
 import com.suhininalex.suffixtree.Node;
-import org.accula.api.detector.psi.Token;
 import org.accula.api.detector.psi.TraverseUtils;
 
 import java.util.HashMap;
